♻️ move search-options types into one file
