+++ apps/web/ui/global-hotkey-provider/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 import * as React from "react";
 import { useParams, useRouter } from "next/navigation";
-import type { OptionGroups } from "~/lib/shared-utils";
+import type { OptionGroups } from "~/lib/search-options";
 
 type GlobalHotkeyContextType = {
   isSearchModalOpen: boolean;
