+++ apps/web/ui/search/search-modal.tsx
@@ -22,7 +22,7 @@ import Image from "next/image";
 import { useFilteredOptionGroup } from "./use-filtered-option-group";
 
 // types
-import type { SearchOption, OptionGroups } from "~/lib/shared-utils";
+import type { SearchOption, OptionGroups } from "~/lib/search-options";
 interface Props {
   defaultNetwork: {
     value: SearchOption;
